refactor/ copyCanvas 관심사 분리
